Moved atmospherics
